move _write_pem to base client
