+++ python/hsfs/client/base.py
@@ -14,8 +14,11 @@
 #   limitations under the License.
 #
 
+import base64
 import os
+import textwrap
 import furl
+from pathlib import Path
 from abc import ABC, abstractmethod
 
 import requests
@@ -23,6 +26,11 @@ import urllib3
 
 from hsfs.client import exceptions, auth
 from hsfs.decorators import connected
+
+try:
+    import jks
+except ImportError:
+    pass
 
 
 urllib3.disable_warnings(urllib3.exceptions.SecurityWarning)
@@ -181,3 +189,91 @@ class Client(ABC):
     def _close(self):
         """Closes a client. Can be implemented for clean up purposes, not mandatory."""
         self._connected = False
+
+    def _write_pem(
+        self, keystore_path, keystore_pw, truststore_path, truststore_pw, prefix
+    ):
+        ks = jks.KeyStore.load(Path(keystore_path), keystore_pw, try_decrypt_keys=True)
+        ts = jks.KeyStore.load(
+            Path(truststore_path), truststore_pw, try_decrypt_keys=True
+        )
+
+        ca_chain_path = os.path.join("/tmp", f"{prefix}_ca_chain.pem")
+        self._write_ca_chain(ks, ts, ca_chain_path)
+
+        client_cert_path = os.path.join("/tmp", f"{prefix}_client_cert.pem")
+        self._write_client_cert(ks, client_cert_path)
+
+        client_key_path = os.path.join("/tmp", f"{prefix}_client_key.pem")
+        self._write_client_key(ks, client_key_path)
+
+        return ca_chain_path, client_cert_path, client_key_path
+
+    def _write_ca_chain(self, ks, ts, ca_chain_path):
+        """
+        Converts JKS keystore and truststore file into ca chain PEM to be compatible with Python libraries
+        """
+        if os.path.exists(ca_chain_path):
+            return
+
+        ca_chain = ""
+        for store in [ks, ts]:
+            for alias, c in store.certs.items():
+                ca_chain = ca_chain + self._bytes_to_pem_str(c.cert, "CERTIFICATE")
+
+        with Path(ca_chain_path).open("w") as f:
+            f.write(ca_chain)
+
+    def _write_client_cert(self, ks, client_cert_path):
+        """
+        Converts JKS keystore file into client cert PEM to be compatible with Python libraries
+        """
+        if os.path.exists(client_cert_path):
+            return
+
+        client_cert = ""
+        for alias, pk in ks.private_keys.items():
+            for c in pk.cert_chain:
+                client_cert = client_cert + self._bytes_to_pem_str(c[1], "CERTIFICATE")
+
+        with Path(client_cert_path).open("w") as f:
+            f.write(client_cert)
+
+    def _write_client_key(self, ks, client_key_path):
+        """
+        Converts JKS keystore file into client key PEM to be compatible with Python libraries
+        """
+        if os.path.exists(client_key_path):
+            return
+
+        client_key = ""
+        for alias, pk in ks.private_keys.items():
+            client_key = client_key + self._bytes_to_pem_str(
+                pk.pkey_pkcs8, "PRIVATE KEY"
+            )
+
+        with Path(client_key_path).open("w") as f:
+            f.write(client_key)
+
+    def _bytes_to_pem_str(self, der_bytes, pem_type):
+        """
+        Utility function for creating PEM files
+
+        Args:
+            der_bytes: DER encoded bytes
+            pem_type: type of PEM, e.g Certificate, Private key, or RSA private key
+
+        Returns:
+            PEM String for a DER-encoded certificate or private key
+        """
+        pem_str = ""
+        pem_str = pem_str + "-----BEGIN {}-----".format(pem_type) + "\n"
+        pem_str = (
+            pem_str
+            + "\r\n".join(
+                textwrap.wrap(base64.b64encode(der_bytes).decode("ascii"), 64)
+            )
+            + "\n"
+        )
+        pem_str = pem_str + "-----END {}-----".format(pem_type) + "\n"
+        return pem_str
